rename(upload):remove-v2-suffix
