+++ web/packages/gallery/components/upload-progress/MinimizedUploadProgress.tsx
