+++ web/packages/gallery/components/upload-progress/StopUploadConfirmationDialog.tsx
