+++ web/packages/gallery/components/upload-progress/UploadProgressDetails.tsx
