+++ web/packages/gallery/components/upload-progress/helpers.ts
