+++ web/packages/gallery/components/upload-progress/useMinimizedUploadDrag.ts
